+++ docs/changes.md
@@ -3,7 +3,25 @@ title: Tenanti Change Log
 
 ---
 
+## Version 3.3 {#v3-3}
+
+### v3.3.0 {#v3-3-0}
+
+* Update support to Laravel Framework v5.3.
+* Remove deprecated `Orchestra\Tenanti\Migrator\Factory::asDefaultDatabase()` method, use `asDefaultConnection()` instead.
+* Remove deprecated `Orchestra\Tenanti\TenantiManager::setupMultiDatabase()` method, use `connection()` instead.
+
 ## Version 3.2 {#v3-2}
+
+### v3.2.9 {#v3-2-9}
+
+* Add `php artisan tenanti:tinker {driver} {id}` to test working on individual tenant.
+* Append `_{id}` to database only when working on none shared database environment.
+
+### v3.2.8 {#v3-2-8}
+
+* Add `Orchestra\Tenanti\Migrator\Operation::resolveConnection()` method.
+* Rename `Orchestra\Tenanti\Migrator\Operation::resolveDatabaseConnection()` to `Orchestra\Tenanti\Migrator\Operation::asConnection()`.
 
 ### v3.2.7 {#v3-2-7}